+++ lib/eventum/class.custom_field.php
@@ -344,7 +344,7 @@ class Custom_Field
 
         $prj_id = Auth::getCurrentProject();
 
-        return Link_Filter::processText($prj_id, Misc::htmlentities($value), 'link', true);
+        return Link_Filter::processText($prj_id, Misc::htmlentities($value), true);
     }
 
     /**