+++ src/EmailHelper.php
@@ -13,13 +13,9 @@
 
 namespace Eventum;
 
-use Auth;
-use EmailReplyParser\EmailReplyParser;
 use EmailReplyParser\Fragment;
 use EmailReplyParser\Parser\EmailParser;
-use Eventum\Db\Doctrine;
 use Link_Filter;
-use Misc;
 
 class EmailHelper
 {
@@ -34,39 +30,8 @@ class EmailHelper
      */
     public static function formatEmail(string $text): string
     {
-        static $enabled;
-
-        if ($enabled === null) {
-            $usr_id = Auth::getUserID();
-            $prefs = Doctrine::getUserPreferenceRepository()->findOrCreate($usr_id);
-            $enabled = $prefs->isMarkdownEnabled();
-        }
-
-        $text = self::collapseReplies($text, $enabled);
-
-        if (!Link_Filter::markdownEnabled()) {
-            $text = nl2br($text);
-            $text = Link_Filter::activateLinks($text);
-        }
-
-        return $text;
-    }
-
-    /**
-     * Collapse email replies and signatures into expandable block
-     */
-    private static function collapseReplies(string $text, bool $enabled): string
-    {
-        if (!$enabled) {
-            return Misc::highlightQuotedReply($text);
-        }
-
-        $wrapText = function ($text) {
-            if (Link_Filter::markdownEnabled()) {
-                $text = Link_Filter::markdownFormat($text);
-            } else {
-                $text = Misc::highlightQuotedReply($text);
-            }
+        $wrapText = static function ($text) {
+            $text = Link_Filter::markdownFormat($text);
 
             return
                 '<div><span class="toggle-trimmed-email"><a href="#">…</a></span>'
@@ -84,16 +49,12 @@ class EmailHelper
 
         // wrap quoted and signature blocks with a div
         $blocks = array_map(
-            function (Fragment $fragment) use ($wrapText) {
+            static function (Fragment $fragment) use ($wrapText) {
                 if ($fragment->isQuoted() || $fragment->isSignature()) {
                     return $wrapText($fragment);
                 }
 
-                if (Link_Filter::markdownEnabled()) {
-                    return Link_Filter::markdownFormat((string)$fragment);
-                }
-
-                return htmlspecialchars($fragment);
+                return Link_Filter::markdownFormat((string)$fragment);
             }, $email->getFragments()
         );
         $text = rtrim(implode("\n", $blocks));